external-api: types: admin: add admin api types
