+++ external-api/src/types/mod.rs
@@ -1,10 +1,12 @@
 //! API types for the relayer's websocket and HTTP APIs
 
+mod admin;
 mod api_network_info;
 mod api_order_book;
 mod api_task_history;
 mod api_wallet;
 
+pub use admin::*;
 pub use api_network_info::*;
 pub use api_order_book::*;
 pub use api_task_history::*;
